+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/transforms/RexConverter.kt
@@ -25,7 +25,6 @@ import org.partiql.ast.Type
 import org.partiql.ast.visitor.AstBaseVisitor
 import org.partiql.planner.internal.Env
 import org.partiql.planner.internal.ir.Identifier
-import org.partiql.planner.internal.ir.Rel
 import org.partiql.planner.internal.ir.Rex
 import org.partiql.planner.internal.ir.builder.plan
 import org.partiql.planner.internal.ir.fnUnresolved
@@ -44,7 +43,9 @@ import org.partiql.planner.internal.ir.rexOpStruct
 import org.partiql.planner.internal.ir.rexOpStructField
 import org.partiql.planner.internal.ir.rexOpSubquery
 import org.partiql.planner.internal.ir.rexOpTupleUnion
+import org.partiql.planner.internal.ir.rexOpVarResolved
 import org.partiql.planner.internal.ir.rexOpVarUnresolved
+import org.partiql.planner.internal.transforms.RelConverter.nil
 import org.partiql.planner.internal.typer.toNonNullStaticType
 import org.partiql.planner.internal.typer.toStaticType
 import org.partiql.types.StaticType
@@ -635,38 +636,37 @@ internal object RexConverter {
         override fun visitExprSFW(node: Expr.SFW, context: Env): Rex = RelConverter.apply(node, context)
 
         override fun visitExprBagOp(node: Expr.BagOp, ctx: Env): Rex {
-            val lhs = Rel(
-                type = Rel.Type(listOf(Rel.Binding("_0", StaticType.ANY)), props = emptySet()),
-                op = Rel.Op.Scan(visitExpr(node.lhs, ctx))
-            )
-            val rhs = Rel(
-                type = Rel.Type(listOf(Rel.Binding("_1", StaticType.ANY)), props = emptySet()),
-                op = Rel.Op.Scan(visitExpr(node.rhs, ctx))
-            )
-            val quantifier = when (node.type.setq) {
-                SetQuantifier.ALL -> org.partiql.planner.internal.ir.SetQuantifier.ALL
-                null, SetQuantifier.DISTINCT -> org.partiql.planner.internal.ir.SetQuantifier.DISTINCT
-            }
-            val isOuter = node.outer == true
-            val op = when (node.type.type) {
-                SetOp.Type.UNION -> Rel.Op.Union(quantifier, lhs, rhs, isOuter)
-                SetOp.Type.EXCEPT -> Rel.Op.Except(quantifier, lhs, rhs, isOuter)
-                SetOp.Type.INTERSECT -> Rel.Op.Intersect(quantifier, lhs, rhs, isOuter)
-            }
-            val rel = Rel(
-                type = Rel.Type(listOf(Rel.Binding("_0", StaticType.ANY)), props = emptySet()),
-                op = op
-            )
-            return Rex(
-                type = StaticType.ANY,
-                op = Rex.Op.Select(
-                    constructor = Rex(
-                        StaticType.ANY,
-                        Rex.Op.Var.Unresolved(Identifier.Symbol("_0", Identifier.CaseSensitivity.SENSITIVE), Rex.Op.Var.Scope.LOCAL)
-                    ),
-                    rel = rel
+            if (node.outer == true) {
+                // PartiQL bag op; create bag op rex
+                val lhs = visitExpr(node.lhs, ctx)
+                val rhs = visitExpr(node.rhs, ctx)
+                val setq = when (node.type.setq) {
+                    SetQuantifier.ALL -> org.partiql.planner.internal.ir.SetQuantifier.ALL
+                    null, SetQuantifier.DISTINCT -> org.partiql.planner.internal.ir.SetQuantifier.DISTINCT
+                }
+                val op = when (node.type.type) {
+                    SetOp.Type.UNION -> Rex.Op.Union(setq, lhs, rhs)
+                    SetOp.Type.EXCEPT -> Rex.Op.Except(setq, lhs, rhs)
+                    SetOp.Type.INTERSECT -> Rex.Op.Intersect(setq, lhs, rhs)
+                }
+                return Rex(
+                    type = StaticType.ANY,
+                    op = op
                 )
-            )
+            } else {
+                // SQL set op; create set op rel
+                val rel = node.accept(RelConverter.ToRel(ctx), nil)
+                return Rex(
+                    type = StaticType.ANY,
+                    op = Rex.Op.Select(
+                        constructor = Rex(
+                            StaticType.ANY,
+                            rexOpVarResolved(0)
+                        ),
+                        rel = rel
+                    )
+                )
+            }
         }
 
         // Helpers